chore: updated serverless framework lambdas
refs 20261
