+++ packages/aws-lambda/.eslintignore
@@ -1,1 +1,2 @@
 **/node_modules/*
+lambdas/serverless-framework/with-serverless-webpack/*
