+++ packages/aws-lambda/lambdas/serverless-framework/.gitignore
@@ -1,0 +1,2 @@
+**/.serverless/
+**/.serverless-offline/
