+++ packages/aws-lambda/lambdas/serverless-framework/with-layer/index.js
@@ -1,0 +1,12 @@
+/*
+ * (c) Copyright IBM Corp. 2022
+ */
+
+'use strict';
+
+exports.handler = async () => {
+  return {
+    statusCode: 200,
+    body: 'Hello from Lambda layer'
+  };
+};
