+++ packages/aws-lambda/lambdas/serverless-framework/with-layer/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "nodejs-team-serverless-with-layer",
+  "description": "A lambda function built with serverless & Instana layer.",
+  "version": "1.0.0",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/instana/nodejs.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "serverless package",
+    "deploy": "serverless deploy",
+    "deploy:nobuild": "serverless deploy --package ./.serverless"
+  },
+  "author": "Bastian Krol <bastian.krol@instana.com>",
+  "license": "MIT",
+  "devDependencies": {
+    "serverless": "3.21.0"
+  }
+}
