+++ packages/aws-lambda/lambdas/serverless-framework/with-serverless-offline/serverless.yml
@@ -1,0 +1,24 @@
+service: nodejs-team-serverless-2
+
+provider:
+  name: aws
+  runtime: nodejs16.x
+  stage: dev
+  region: us-east-2
+
+plugins:
+  - serverless-offline
+
+custom:
+  serverless-offline:
+    useDocker: true
+
+functions:
+  with-serverless-offline:
+    environment:
+      NODE_PATH: $NODE_PATH:/opt/nodejs/node_modules
+      INSTANA_ENDPOINT_URL: ${env:INSTANA_ENDPOINT_URL}
+      INSTANA_AGENT_KEY: ${env:INSTANA_AGENT_KEY}
+    layers:
+      - arn:aws:lambda:us-east-2:410797082306:layer:instana-nodejs:110
+    handler: instana-aws-lambda-auto-wrap.handler
