+++ packages/aws-lambda/lambdas/serverless-framework/with-serverless-webpack/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "serverless-webpack-test",
-  "description": "A lambda function built with serverless and webpack",
+  "name": "nodejs-team-serverless-with-serverless-webpack",
+  "description": "A lambda function built with serverless & serverless webpack",
   "version": "1.0.0",
   "main": "index.js",
   "repository": {
@@ -10,8 +10,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "serverless package",
-    "deploy-initial": "sls deploy",
-    "deploy": "sls deploy function -f serverless-webpack-test"
+    "deploy": "serverless deploy",
+    "deploy:nobuild": "serverless deploy --no-build"
   },
   "author": "Bastian Krol <bastian.krol@instana.com>",
   "license": "MIT",
@@ -19,9 +19,9 @@
     "lodash": "^4.17.15"
   },
   "devDependencies": {
-    "eslint-plugin-import": "^2.18.2",
-    "serverless": "^1.57.0",
-    "serverless-bundle": "^1.2.2",
-    "serverless-offline": "^5.3.3"
+    "babel-loader": "^8.2.5",
+    "serverless": "3.21.0",
+    "serverless-webpack": "^5.8.0",
+    "webpack-node-externals": "^3.0.0"
   }
 }
