--- packages/aws-lambda/lambdas/serverless/.eslintrc.js
@@ -1,15 +1,0 @@
-/*
- * (c) Copyright IBM Corp. 2021
- * (c) Copyright Instana Inc. and contributors 2019
- */
-
-module.exports = {
-  parserOptions: {
-    sourceType: 'module'
-  },
-
-  rules: {
-    'no-console': 'off',
-    'import/prefer-default-export': 'off'
-  }
-};
