--- packages/aws-lambda/lambdas/serverless/.gitignore
@@ -1,1 +1,0 @@
-.serverless/
