--- packages/aws-lambda/lambdas/serverless/README.md
@@ -1,5 +1,0 @@
-# Serverless/Webpack Test Lambda
-
-* Build package: `npm run build` (or `serverless package`)
-* Build and deploy (first deployment): `npm run deploy-initial` (or `sls deploy`)
-* Build and deploy (a bit faster, after initial deploy): `npm run deploy` (or `sls deploy function -f serverless-webpack-test`)
