--- packages/aws-lambda/lambdas/serverless/serverless.yml
@@ -1,21 +1,0 @@
-service: serverless-webpack-test
-
-package:
-  individually: true
-
-plugins:
-  - serverless-bundle # package with Webpack
-  - serverless-offline
-
-provider:
-  name: aws
-  runtime: nodejs10.x
-  stage: dev
-  region: us-east-2
-
-functions:
-  serverless-webpack-test:
-    # Actually this should be instana-aws-lambda-auto-wrap.handler but serverless complains.
-    # Needs to be corrected manually after deployment for now.
-    handler: index.webpackHandler
-
